src/ts/2024/day23/solution.spec.ts
import { getExampleInput, getInput } from "@aoc/common";
import { Solution } from "./solution";

describe("Day 23", () => {
  let solution: Solution;

  beforeEach(() => {
    solution = new Solution();
  });

  describe("part 1", () => {
    test("part 1 example", () => {
      const input = getExampleInput(__dirname);
      expect(solution.part1(input)).toBe(7);
    });
    test("part 1 input", () => {
      const input = getInput(__dirname);
      expect(solution.part1(input)).toBe(1098);
    });
  });

  describe("part 2", () => {
    test("part 2 example", () => {
      const input = getExampleInput(__dirname);
      expect(solution.part2(input)).toBe("co,de,ka,ta");
    });
    test("part 2 input", () => {
      const input = getInput(__dirname);
      expect(solution.part2(input)).toBe(
        "ar,ep,ih,ju,jx,le,ol,pk,pm,pp,xf,yu,zg"
      );
    });
  });
});
